update: check TFServing before start

+++ bin/_start_service_separate.py
@@ -84,16 +84,19 @@ def start_server(serv_name):
 
 def start_tf_serving(serv_name):
     serv_params = CONFIG_TFSERVING[serv_name]
-    cmd = f"""
-    docker run -d --rm -p {serv_params.docker_port}:8501 \
-        --name {serv_params.name} \
-        -v "{serv_params.pb_path}:/models/{serv_params.name}" \
-        -e MODEL_NAME={serv_params.name} \
-        -t tensorflow/serving > {serv_params.logfile}  &
-    """.strip()
-    status, output = subprocess.getstatusoutput(cmd)
-    zprint(">>> 启动内部TFServing服务 {} 于端口 {} ".format(serv_name, serv_params.docker_port))
-    zprint(">>> {}: subprocess status is: ' {} ', output is: ' {} '".format("SUCCESS" if status == 0 else "FAIL", status, output))
+    status, output = subprocess.getstatusoutput("docker ps -a | awk '{print $NF}'")
+    if serv_params.name not in output.split("\n"):
+        zprint(">>> TFServing(docker) 未开启，将开启服务 {} 于端口 {}".format(serv_name, serv_params.docker_port))
+        start_cmd = f"""
+            docker run -d --rm -p {serv_params.docker_port}:8501 \
+                --name {serv_params.name} \
+                -v "{serv_params.pb_path}:/models/{serv_params.name}" \
+                -e MODEL_NAME={serv_params.name} \
+                -t tensorflow/serving > {serv_params.logfile}  &
+            """.strip()
+        status, output = subprocess.getstatusoutput(start_cmd)
+        zprint(">>> {}: subprocess status is: ' {} ', output is: ' {} '".format("SUCCESS" if status == 0 else "FAIL", status, output))
+    zprint(">>> 内部TFServing服务 {} 已启动于端口 {} ".format(serv_name, serv_params.docker_port))
 
 
 if SERVICE == "all":
